fix: move logout action
